refactor(builders): share the five identical write resolvers between pg and sqlite
generateInsertArray, generateInsertSingle, generateUpsert, generateUpdate and
generateDelete were duplicated verbatim between pg.ts and sqlite.ts — measured
at 94.7%-97.9% identical, with the differences being only the dialect spelling
of the `db` and `table` parameter types, the primary-key lookup, and one
explicit generic argument.

They now live in src/util/builders/write-resolvers.ts behind
buildWriteResolvers(primaryKeyPropNames), which each dialect calls once and
destructures, so all 22 call sites are unchanged. The bodies are pg.ts's
verbatim. MySQL is deliberately left alone: its mutations return no rows, so it
re-selects after every write and its resolvers have a genuinely different shape.

pg.ts 2086 -> 1340 lines, sqlite.ts 1987 -> 1239. dist/index.d.ts is
byte-identical before and after.
